refactor(#244): ScheduleView 컴포넌트 분리
